requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.2.5
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
greenlet==3.0.1
gunicorn==21.2.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pillow==10.2.0
SQLAlchemy==2.0.22
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
